Revised Dg class and subclass, flash point and packing group properties of Dg. Made necessary changes to other classes to keep the program operational.

+++ EasyJob Pro DG.Model/Cargo/Container.cs
@@ -16,8 +16,8 @@
         #region IReefer
         // -------------- IReefer ---------------------------------------------------
 
-        public double SetTemperature { get; set; }
-        public double LoadTemperature { get; set; }
+        public decimal SetTemperature { get; set; }
+        public decimal LoadTemperature { get; set; }
         public string Commodity { get; set; }
         public string VentSetting { get; set; }
         public string ReeferSpecial { get; set; }
